+++ src/tests/Recyclarr.Cli.Tests/Pipelines/QualityProfile/PipelinePhases/QualityProfileTransactionPhaseTest.cs
@@ -151,7 +151,7 @@ public class QualityProfileTransactionPhaseTest
     }
 
     [Test, AutoMockData]
-    public void Reset_scores(
+    public void Reset_scores_with_reset_unmatched_true(
         QualityProfileTransactionPhase sut)
     {
         var guideData = new[]
@@ -190,6 +190,50 @@ public class QualityProfileTransactionPhaseTest
             {
                 NewQp.UpdatedScore("quality1", 200, 0, FormatScoreUpdateReason.Reset),
                 NewQp.UpdatedScore("quality2", 300, 0, FormatScoreUpdateReason.Reset),
+                NewQp.UpdatedScore("quality3", 0, 100, FormatScoreUpdateReason.New),
+                NewQp.UpdatedScore("quality4", 0, 500, FormatScoreUpdateReason.New)
+            }, o => o.Excluding(x => x.Dto.Format));
+    }
+
+    [Test, AutoMockData]
+    public void Reset_scores_with_reset_unmatched_false(QualityProfileTransactionPhase sut)
+    {
+        var guideData = new[]
+        {
+            NewQp.Processed("profile1", false, ("quality3", "id3", 3, 100), ("quality4", "id4", 4, 500))
+        };
+
+        var serviceData = new[]
+        {
+            new QualityProfileDto
+            {
+                Name = "profile1",
+                FormatItems = new[]
+                {
+                    new ProfileFormatItemDto
+                    {
+                        Name = "quality1",
+                        Format = 1,
+                        Score = 200
+                    },
+                    new ProfileFormatItemDto
+                    {
+                        Name = "quality2",
+                        Format = 2,
+                        Score = 300
+                    }
+                }
+            }
+        };
+
+        var result = sut.Execute(guideData, serviceData);
+
+        result.UpdatedProfiles.Should()
+            .ContainSingle().Which.UpdatedScores.Should()
+            .BeEquivalentTo(new[]
+            {
+                NewQp.UpdatedScore("quality1", 200, 200, FormatScoreUpdateReason.Reset),
+                NewQp.UpdatedScore("quality2", 300, 300, FormatScoreUpdateReason.Reset),
                 NewQp.UpdatedScore("quality3", 0, 100, FormatScoreUpdateReason.New),
                 NewQp.UpdatedScore("quality4", 0, 500, FormatScoreUpdateReason.New)
             }, o => o.Excluding(x => x.Dto.Format));